requirements.txt
anyio
httpx
httpx-sse
sse-starlette
starlette
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
typing-extensions
typing-inspection
jsonschema
openapi-spec-validator
prance
structlog
pyyaml
pytest
pytest-asyncio
respx
ruff
